+++ src/types/product.ts
@@ -1,0 +1,12 @@
+export type Product = {
+  pointProductId: number
+  pointProductName: string
+  thumbnailUrl: string
+  pointPrice: number
+  sellingStatus: string
+}
+
+export type ProdcutDetailType = Product & {
+  description: string
+  stockQuantity: number
+}
